Unsure how to proceed; user cases are contradictory
